requirements.txt
django
django-cors-headers
django-ninja
django-ninja-jwt[crypto]
rav